+++ eCar/UI/ecar_admin/lib/screens/vehicle_screen.dart
@@ -5,6 +5,7 @@ import 'package:ecar_admin/models/search_result.dart';
 import 'package:ecar_admin/providers/vehicle_provider.dart';
 import 'package:ecar_admin/screens/master_screen.dart';
 import 'package:ecar_admin/screens/vehicle_details_screen.dart';
+import 'package:ecar_admin/utils/alert_helpers.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:ecar_admin/utils/string_helpers.dart' as help;
@@ -226,11 +227,16 @@ class _VehicleScreenState extends State<VehicleScreen> {
                               DataCell(
                                 ElevatedButton(
                                   onPressed: () async {
-                                    provider.delete(e.id);
-                                    await Future.delayed(
-                                        const Duration(seconds: 1));
-                                    result = await provider.get();
-                                    setState(() {});
+                                    bool? confirmDelete =
+                                        await AlertHelpers.deleteConfirmation(
+                                            context);
+                                    if (confirmDelete == true) {
+                                      provider.delete(e.id);
+                                      await Future.delayed(
+                                          const Duration(seconds: 1));
+                                      result = await provider.get();
+                                      setState(() {});
+                                    }
                                   },
                                   style: ButtonStyle(
                                     backgroundColor: MaterialStateProperty.all(
